+++ cynic-querygen/tests/starwars-tests.rs
@@ -25,3 +25,4 @@ test_query_file!(bare_selection_sets, "bare-selection-set.graphql");
 test_query_file!(multiple_queries, "multiple-queries.graphql");
 test_query_file!(fragment_spreads, "fragment-spreads.graphql");
 test_query_file!(aliases, "aliases.graphql");
+test_query_file!(float, "float.graphql");