+++ CHANGELOG.md
@@ -21,6 +21,7 @@ all APIs might be changed.
 
 - Querygen no longer duplicates variable names in generated `ArgumentStruct`s
   when an argument is used twice in a query.
+- Fixed support for `Float` scalars in the generator.
 
 ## v0.15.0 - 2021-09-23